+++ test/opf-var.jl
@@ -187,6 +187,208 @@ end
 
 
 
+@testset "test opf with swtiches" begin
+
+    @testset "test ac opf" begin
+        @testset "5-bus sw case" begin
+            result = PowerModels._run_sw_opf("../test/data/matpower/case5_sw.m", ACPPowerModel, ipopt_solver)
+
+            @test result["termination_status"] == LOCALLY_SOLVED
+            @test isapprox(result["objective"], 16641.2; atol = 1e0)
+            @test isapprox(result["solution"]["switch"]["1"]["psw"], 3.051, atol=1e-2)
+            @test isapprox(result["solution"]["switch"]["1"]["qsw"], 0.885, atol=1e-2)
+            @test isapprox(result["solution"]["switch"]["2"]["psw"], 0.000, atol=1e-3)
+            @test isapprox(result["solution"]["switch"]["2"]["qsw"], 0.000, atol=1e-3)
+            @test isnan(result["solution"]["switch"]["3"]["psw"])
+            @test isnan(result["solution"]["switch"]["3"]["qsw"])
+        end
+
+        @testset "5-bus sw nb case" begin
+            result = PowerModels._run_sw_opf("../test/data/matpower/case5_sw_nb.m", ACPPowerModel, ipopt_solver)
+
+            @test result["termination_status"] == LOCALLY_SOLVED
+            @test isapprox(result["objective"], 17915.3; atol = 1e0)
+        end
+    end
+
+    @testset "test dc opf" begin
+        @testset "5-bus sw case" begin
+            result = PowerModels._run_sw_opf("../test/data/matpower/case5_sw.m", DCPPowerModel, cbc_solver)
+
+            @test result["termination_status"] == OPTIMAL
+            @test isapprox(result["objective"], 16554.7; atol = 1e0)
+            @test isapprox(result["solution"]["switch"]["1"]["psw"], 3.050, atol=1e-2)
+            @test isnan(result["solution"]["switch"]["1"]["qsw"])
+            @test isapprox(result["solution"]["switch"]["2"]["psw"], 0.000, atol=1e-3)
+            @test isnan(result["solution"]["switch"]["2"]["qsw"])
+        end
+
+        @testset "5-bus sw nb case" begin
+            result = PowerModels._run_sw_opf("../test/data/matpower/case5_sw_nb.m", DCPPowerModel, cbc_solver)
+
+            @test result["termination_status"] == OPTIMAL
+            @test isapprox(result["objective"], 17751.3; atol = 1e0)
+        end
+    end
+
+    @testset "test soc opf" begin
+        @testset "5-bus sw case" begin
+            result = PowerModels._run_sw_opf("../test/data/matpower/case5_sw.m", SOCWRPowerModel, ipopt_solver)
+
+            @test result["termination_status"] == LOCALLY_SOLVED
+            @test isapprox(result["objective"], 15110.0; atol = 1e0)
+            @test isapprox(result["solution"]["switch"]["1"]["psw"], 3.048, atol=1e-2)
+            @test isapprox(result["solution"]["switch"]["1"]["qsw"], 0.889, atol=1e-2)
+            @test isapprox(result["solution"]["switch"]["2"]["psw"], 0.000, atol=1e-3)
+            @test isapprox(result["solution"]["switch"]["2"]["qsw"], 0.000, atol=1e-3)
+        end
+
+        @testset "5-bus sw nb case" begin
+            result = PowerModels._run_sw_opf("../test/data/matpower/case5_sw_nb.m", SOCWRPowerModel, ipopt_solver)
+
+            @test result["termination_status"] == LOCALLY_SOLVED
+            @test isapprox(result["objective"], 15192.8; atol = 1e0)
+        end
+    end
+
+end
+
+
+@testset "test oswpf" begin
+
+    @testset "test ac oswpf" begin
+        @testset "5-bus sw case" begin
+            result = PowerModels._run_oswpf("../test/data/matpower/case5_sw.m", ACPPowerModel, juniper_solver)
+
+            @test result["termination_status"] == LOCALLY_SOLVED
+            @test isapprox(result["objective"], 15053.6; atol = 1e0)
+            @test isapprox(result["solution"]["switch"]["1"]["psw"],  5.468, atol=1e-2)
+            @test isapprox(result["solution"]["switch"]["1"]["qsw"], -0.836, atol=1e-2)
+            @test isapprox(result["solution"]["switch"]["2"]["psw"], -2.426, atol=1e-3)
+            @test isapprox(result["solution"]["switch"]["2"]["qsw"],  1.736, atol=1e-3)
+
+            @test isapprox(result["solution"]["switch"]["1"]["status"], 1.00, atol=1e-3)
+            @test isapprox(result["solution"]["switch"]["2"]["status"], 1.00, atol=1e-3)
+        end
+
+        @testset "5-bus sw nb case" begin
+            result = PowerModels._run_oswpf("../test/data/matpower/case5_sw_nb.m", ACPPowerModel, juniper_solver)
+
+            @test result["termination_status"] == LOCALLY_SOLVED
+            @test isapprox(result["objective"], 16674.8; atol = 1e0)
+
+            switch_status_total = sum(switch["status"] for (i,switch) in result["solution"]["switch"])
+            @test isapprox(switch_status_total, 12.00, atol=1e-4) # two swtiches off
+        end
+    end
+
+    @testset "test dc oswpf" begin
+        @testset "5-bus sw case" begin
+            result = PowerModels._run_oswpf("../test/data/matpower/case5_sw.m", DCPPowerModel, cbc_solver)
+
+            @test result["termination_status"] == OPTIMAL
+            @test isapprox(result["objective"], 15054.1; atol = 1e0)
+
+            @test isapprox(result["solution"]["switch"]["1"]["psw"], 5.603, atol=1e-2)
+            @test isnan(result["solution"]["switch"]["1"]["qsw"])
+            @test isapprox(result["solution"]["switch"]["2"]["psw"], -2.553, atol=1e-3)
+            @test isnan(result["solution"]["switch"]["2"]["qsw"])
+
+            @test isapprox(result["solution"]["switch"]["1"]["status"], 1.00, atol=1e-3)
+            @test isapprox(result["solution"]["switch"]["2"]["status"], 1.00, atol=1e-3)
+        end
+
+        @testset "5-bus sw nb case" begin
+            result = PowerModels._run_oswpf("../test/data/matpower/case5_sw_nb.m", DCPPowerModel, cbc_solver)
+
+            @test result["termination_status"] == OPTIMAL
+            @test isapprox(result["objective"], 15141.2; atol = 1e0)
+
+            switch_status_total = sum(switch["status"] for (i,switch) in result["solution"]["switch"])
+            @test switch_status_total <= 12.000 && switch_status_total >= 10.000 # two to four swtiches off
+        end
+    end
+
+    @testset "test soc oswpf" begin
+        @testset "5-bus sw case" begin
+            result = PowerModels._run_oswpf("../test/data/matpower/case5_sw.m", SOCWRPowerModel, juniper_solver)
+
+            @test result["termination_status"] == LOCALLY_SOLVED
+            @test isapprox(result["objective"], 15053.6; atol = 1e0)
+            @test isapprox(result["solution"]["switch"]["1"]["psw"],  5.469, atol=1e-2)
+            @test isapprox(result["solution"]["switch"]["1"]["qsw"], -0.809, atol=1e-2)
+            @test isapprox(result["solution"]["switch"]["2"]["psw"], -2.426, atol=1e-3)
+            @test isapprox(result["solution"]["switch"]["2"]["qsw"],  1.710, atol=1e-3)
+
+            @test isapprox(result["solution"]["switch"]["1"]["status"], 1.00, atol=1e-3)
+            @test isapprox(result["solution"]["switch"]["2"]["status"], 1.00, atol=1e-3)
+        end
+
+        @testset "5-bus sw nb case" begin
+            result = PowerModels._run_oswpf("../test/data/matpower/case5_sw_nb.m", SOCWRPowerModel, juniper_solver)
+
+            @test result["termination_status"] == LOCALLY_SOLVED
+            @test isapprox(result["objective"], 15175.7; atol = 1e0)
+
+            switch_status_total = sum(switch["status"] for (i,switch) in result["solution"]["switch"])
+            @test isapprox(switch_status_total, 13.00, atol=1e-4) # one swtich off
+        end
+    end
+
+end
+
+
+
+@testset "test oswpf node-breaker" begin
+
+    @testset "test ac oswpf node-breaker" begin
+        @testset "5-bus sw nb case" begin
+            result = PowerModels._run_oswpf_nb("../test/data/matpower/case5_sw_nb.m", ACPPowerModel, juniper_solver)
+
+            @test result["termination_status"] == LOCALLY_SOLVED
+            @test isapprox(result["objective"], 15350.4; atol = 1e0)
+
+            switch_status_total = sum(switch["status"] for (i,switch) in result["solution"]["switch"])
+            @test isapprox(switch_status_total, 10.00, atol=1e-4) # four swtiches off
+
+            branch_status_total = sum(branch["br_status"] for (i,branch) in result["solution"]["branch"])
+            @test isapprox(branch_status_total, 5.00, atol=1e-4) # two branches off
+        end
+    end
+
+    @testset "test dc oswpf node-breaker" begin
+        @testset "5-bus sw nb case" begin
+            result = PowerModels._run_oswpf_nb("../test/data/matpower/case5_sw_nb.m", DCPPowerModel, cbc_solver)
+
+            @test result["termination_status"] == OPTIMAL
+            @test isapprox(result["objective"], 15141.2; atol = 1e0)
+
+            switch_status_total = sum(switch["status"] for (i,switch) in result["solution"]["switch"])
+            @test switch_status_total <= 13.000 && switch_status_total >= 12.000 # 1 to 2 swtiches off
+
+            branch_status_total = sum(branch["br_status"] for (i,branch) in result["solution"]["branch"])
+            @test isapprox(branch_status_total, 6.00, atol=1e-4) # one branch off
+        end
+    end
+
+    @testset "test soc oswpf node-breaker" begin
+        @testset "5-bus sw nb case" begin
+            result = PowerModels._run_oswpf_nb("../test/data/matpower/case5_sw_nb.m", SOCWRPowerModel, juniper_solver)
+
+            @test result["termination_status"] == LOCALLY_SOLVED
+            @test isapprox(result["objective"], 15175.7; atol = 1e0)
+
+            switch_status_total = sum(switch["status"] for (i,switch) in result["solution"]["switch"])
+            @test isapprox(switch_status_total, 13.00, atol=1e-4) # one swtich off
+
+            branch_status_total = sum(branch["br_status"] for (i,branch) in result["solution"]["branch"])
+            @test isapprox(branch_status_total, 7.00, atol=1e-4) # no branches off
+        end
+    end
+
+end
+
+
 @testset "test storage opf" begin
 
     @testset "test ac polar opf" begin